+++ tests/conftest.py
@@ -26,6 +26,7 @@ def threedi_db(tmpdir_factory):
     db = ThreediDatabase(tmp_sqlite)
     schema = ModelSchema(db)
     schema.upgrade(backup=False, upgrade_spatialite_version=False)
+    schema.set_spatial_indexes()
     return db